+++ packages/ocr-ggml/addon/src/model-interface/DoctrOcrModel.hpp
@@ -10,7 +10,7 @@
 #include <inference-addon-cpp/ModelInterfaces.hpp>
 #include <inference-addon-cpp/RuntimeStats.hpp>
 
-#include "OcrModel.hpp"
+#include "OcrTypes.hpp"
 #include "doctr/StepDoctrDetectionGGML.hpp"
 #include "doctr/StepDoctrRecognitionGGML.hpp"
 #include "easyocr/pipeline/steps.hpp"